refactoring reset route
