+++ src/routes/resetRoutes.ts
@@ -1,0 +1,8 @@
+import { Router } from "express";
+import { reset } from "../controllers/resetController";
+
+const router = Router();
+
+router.post("/reset", reset);
+
+export default router;
